Initial SketchChair Beta Commit
This is the first commit of the SketchChair source code! Currently
SketchChair is in Beta and it's source code may not be suitable for use
in other projects. Whilst I have tried to structure the code in clear
logical manner some parts of the source code remain a little untidy and
undocumented. Over the next few months we will be tidying things up,
fixing bug and refining features based on user feedback.